ajout gestion modification supression d'un element

+++ backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/Thing');
+const fs = require('fs')
 
 exports.createThing = (req, res, next) => {
   const thingObject = JSON.parse(req.body.thing);
@@ -16,15 +17,42 @@ exports.createThing = (req, res, next) => {
 };
 
 exports.modifyThing = (req, res, next) => {
-    Thing.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet modifié !'}))
-    .catch(error => res.status(400).json({ error }));
+    const thingObject = req.file ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl:`${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    delete thingObject.userId;
+
+    Thing.findOne({ _id: req.params.id })
+        .then(thing => {
+            if(thing.userId != req.auth.userId){
+                res.status(400).json({ error })
+            } else {
+
+                Thing.updateOne({_id: req.params.id}, {...thingObject, _id: req.params.id})
+                    .then(() => res.status(200).json({message: 'Objet modifié.'}))
+                    .catch(error => res.status(400).json({error}))
+            }
+        } )
+        .catch(error => res.status(400).json({ error }))
 };
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id })
+    .then( thing => {
+        if(thing.userId != req.auth.userId){
+            res.status(401).json({ message: ' non autorisé.'})
+        } else {
+            const filename = thing.imageUrl.split('/image/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Thing.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({message: 'Objet suprimé.'}))
+                .catch(error => res.status(401).json({error}))
+            })
+        }
+    })
+    .catch(error => res.status(500).json({ error }))
 };
 
 exports.getAllThings = (req, res, next) => {
